Fix type mismatch errors inside the `copy` function are not shown
