+++ kopy-compiler/main/kotlin/com/javiersc/kotlin/kopy/compiler/fir/checker/checkers/expression/ArgumentTypeMismatchTypeChecker.kt
@@ -1,0 +1,51 @@
+package com.javiersc.kotlin.kopy.compiler.fir.checker.checkers.expression
+
+import com.javiersc.kotlin.kopy.compiler.fir.checker.FirKopyError
+import com.javiersc.kotlin.kopy.compiler.fir.utils.isKopyFunctionSetCall
+import org.jetbrains.kotlin.diagnostics.DiagnosticReporter
+import org.jetbrains.kotlin.diagnostics.SourceElementPositioningStrategies
+import org.jetbrains.kotlin.diagnostics.reportOn
+import org.jetbrains.kotlin.fir.FirSession
+import org.jetbrains.kotlin.fir.analysis.checkers.MppCheckerKind
+import org.jetbrains.kotlin.fir.analysis.checkers.context.CheckerContext
+import org.jetbrains.kotlin.fir.analysis.checkers.expression.FirCallChecker
+import org.jetbrains.kotlin.fir.expressions.FirCall
+import org.jetbrains.kotlin.fir.expressions.FirFunctionCall
+import org.jetbrains.kotlin.fir.expressions.arguments
+import org.jetbrains.kotlin.fir.types.ConeKotlinType
+import org.jetbrains.kotlin.fir.types.renderReadable
+import org.jetbrains.kotlin.fir.types.resolvedType
+import org.jetbrains.kotlin.fir.types.typeContext
+import org.jetbrains.kotlin.types.AbstractTypeChecker
+
+internal class ArgumentTypeMismatchTypeChecker(
+    private val session: FirSession,
+) : FirCallChecker(MppCheckerKind.Common) {
+
+    override fun check(expression: FirCall, context: CheckerContext, reporter: DiagnosticReporter) {
+        if (!expression.isKopyFunctionSetCall) return
+        if (expression !is FirFunctionCall) return
+
+        val extensionType: ConeKotlinType? = expression.extensionReceiver?.resolvedType
+        val argumentType: ConeKotlinType? = expression.arguments.firstOrNull()?.resolvedType
+
+        if (extensionType == null || argumentType == null) return
+
+        val isSubtype: Boolean =
+            AbstractTypeChecker.isSubtypeOf(
+                context = session.typeContext,
+                subType = extensionType,
+                superType = argumentType,
+            )
+        if (!isSubtype) {
+            reporter.reportOn(
+                source = expression.arguments.first().source,
+                factory = FirKopyError.ARGUMENT_TYPE_MISMATCH,
+                a = extensionType.renderReadable(),
+                b = argumentType.renderReadable(),
+                context = context,
+                positioningStrategy = SourceElementPositioningStrategies.DEFAULT,
+            )
+        }
+    }
+}
